Change requests
* Reduce the size of the logo on the home page. On the live site you'll
see the nav has a couple of extra items -
https://www.coastpowersports.com.au - and we are going to add another
one 'used bikes' which will push the nav into the logo at the current
size. Perhaps make it a bit smaller but not quite as small as the sub
pages?

* On mobile, can we please show the nav bar at all times? And not have
it disappear on scroll? The client has asked for this and I double
checked and they are still keen.